+++ src/core/secureData.ts
@@ -1,7 +1,6 @@
 import fs from "fs";
 import yaml from "yaml";
 import path from "path";
-import { isArrayBuffer } from "util/types";
 
 interface ObjectArray {
   [key: string]: any;
@@ -263,12 +262,12 @@ export async function decodeSQL(
   }
 }
 
-export async function neutralizer(folderPath: string, info: { dataType: "json" | "yaml" | "sql", secret: string }): Promise<any> {
+export async function neutralizer(folderPath: string, info: { dataType: "json" | "yaml" | "sql", secret: string }): Promise<string[]> {
 
-  const extension = "verse";
   const foundFiles: string[] = [];
 
-  if (!info || info.dataType || info.secret || folderPath) throw new Error("Wrong usage: Please Make sure to rpovde folder path and info object parameter { dataType, secret}.")
+  if (!info || !info.dataType || !info.secret || !folderPath) throw new Error("Wrong usage: Please Make sure to provide folder path and info object parameter { dataType, secret }.");
+
   function searchFiles(currentPath: string): void {
       const files = fs.readdirSync(currentPath);
       files.forEach((file) => {
@@ -277,9 +276,9 @@ export async function neutralizer(folderPath: string, info: { dataType: "json" |
           if (stats.isDirectory()) {
               searchFiles(filePath);
           } else {
-              const fileExtension = path.extname(file).toLowerCase();
-              if (fileExtension === `.${extension}`) {
-                  foundFiles.push(filePath); 
+              const fileExtension = path.extname(file);
+              if (fileExtension === '.verse') {
+                  foundFiles.push(filePath);
               }
           }
       });
@@ -287,31 +286,38 @@ export async function neutralizer(folderPath: string, info: { dataType: "json" |
 
   searchFiles(folderPath);
 
-  if (info.dataType === 'json') {    
-      for (let i = 0; i < foundFiles.length; i++) {
-        const filePath = foundFiles[i];
-        const decodedData: any = await decodeJSON(filePath, info.secret);
-        if (!decodedData || decodedData === null) throw new Error("Failed to decode JSON data.");
-        const jsonData = JSON.stringify(decodedData)
-        await fs.promises.writeFile(filePath, jsonData);
+  for (let i = 0; i < foundFiles.length; i++) {
+      const filePath = foundFiles[i];
+      const fileExtension = path.extname(filePath);
+      const fileBaseName = path.basename(filePath, fileExtension);
+
+      let decodedData: any;
+      let newData: any;
+
+      if (info.dataType === 'json') {
+          decodedData = await decodeJSON(filePath, info.secret);
+          newData = JSON.stringify(decodedData);
+      } else if (info.dataType === 'yaml') {
+          decodedData = await decodeYAML(filePath, info.secret);
+          newData = yaml.stringify(decodedData);
+      } else if (info.dataType === 'sql') {
+          const encodedData = await fs.promises.readFile(filePath, "utf-8");
+          decodedData = await decodeSQL(encodedData, info.secret);
+          newData = decodedData; 
       }
-  } else if (info.dataType === 'yaml') {
-    for (let i = 0; i < foundFiles.length; i++) {
-      const filePath = foundFiles[i];
-      const decodedData: any = await decodeYAML(filePath, info.secret);
-      if (!decodedData || decodedData === null) throw new Error("Failed to decode YAML data.");
-      const yamlData = yaml.stringify(decodedData)
-      await fs.promises.writeFile(filePath, yamlData);
-    } 
-  } else if (info.dataType === 'sql') {
-    for (let i = 0; i < foundFiles.length; i++) {
-      const filePath = foundFiles[i];
-      const encodedData =  await fs.promises.readFile(filePath, "utf-8");
-      const decodedData: any = await decodeSQL(encodedData, info.secret);
-      if (!decodedData || decodedData === null) throw new Error("Failed to decode SQL data.");
-      await fs.promises.writeFile(filePath, decodedData);
-    } 
+
+      if (!decodedData || newData === null) throw new Error(`Failed to decode ${info.dataType} data.`);
+      
+      const newFilePath = path.join(path.dirname(filePath), fileBaseName);
+      
+      if (info.dataType === 'json') {
+          await fs.promises.writeFile(`${newFilePath}.json`, newData);
+      } else if (info.dataType === 'yaml') {
+          await fs.promises.writeFile(`${newFilePath}.yaml`, newData);
+      } else if (info.dataType === 'sql') {
+          await fs.promises.writeFile(`${newFilePath}.sql`, newData);
+      }
   }
-  
+
   return foundFiles;
 }